pyproject.toml
[project]
name = "scalpel-scan"
version = "0.2.0"
description = "SCALPEL-scan: Safety-critical attention head analysis for LLMs (scan component)"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.36",
    "accelerate>=0.26",
    "numpy",
    "datasets>=2.14",
    "pyyaml",
    "matplotlib>=3.7",
]

[project.scripts]
scalpel = "scalpel.cli:main"

[project.optional-dependencies]
judge = ["openai>=1.0"]

[tool.setuptools.packages.find]
where = ["."]
